Added skeleton of init function
The idea is that the full init process will be divided over two
functions - init and config. One creates the repo, the other sets
necessary config. These will be internal functions which will be called
together by a superfunction.

The reason for the division is testing. Smaller functions are easier to
test.